requirements.txt
anyio==4.3.0
bottle==0.12.25
Brotli==1.1.0
certifi==2024.2.2
charset-normalizer==3.3.2
filelock==3.14.0
fire==0.6.0
fsspec==2024.5.0
fugashi==1.3.2
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
huggingface-hub==0.23.1
idna==3.7
jaconv==0.3.4
Jinja2==3.1.4
loguru==0.7.2
manga-ocr==0.1.11
MarkupSafe==2.1.5
mpmath==1.3.0
networkx==3.3
numpy==1.26.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.5.40
nvidia-nvtx-cu12==12.1.105
opencv-python==4.9.0.80
packaging==24.0
pillow==10.3.0
plyer==2.1.0
proxy_tools==0.1.0
pydantic_core==2.18.2
PyExecJS==1.5.1
Pygments==2.18.0
pyperclip==1.8.2
pyreqwest_impersonate==0.4.5
python-dotenv==1.0.1
python-multipart==0.0.9
python-socks==2.4.4
python-telegram-bot==21.2
pywebview==5.1
PyYAML==6.0.1
regex==2024.5.15
requests==2.32.2
safetensors==0.4.3
sentencepiece==0.2.0
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
soupsieve==2.5
sympy==1.12
termcolor==2.4.0
tinycss2==1.3.0
tokenizers==0.19.1
torch==2.3.0
tqdm==4.66.4
transformers==4.41.1
triton==2.3.0
typing_extensions==4.12.0
ujson==5.10.0
unidic-lite==1.0.8
urllib3==2.2.1
uvloop==0.19.0
webencodings==0.5.1
websockets==12.0
Werkzeug==3.0.3
